MLEngine/src/requirements.txt
brotli
scrapy
psycopg2-binary
scrapy-user-agents
joblib